Specify MSRV to be 1.76
Async traits were added in 1.75 and Result::inspect_err in 1.76. Rust
1.76 was released in February 2024.

I've added Cargo.lock so that dependencies do not get randomly updated.
I've also pinned the version for CI, so that we have reproducible
builds.

We use clippy 1.80 because of the generated `meshtastic.rs` for
protobufs. We can deal with this later.